+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Changelog
 
-__5.0.3 - 2026-09-28__
+__5.0.4 - 2026-09-28__
 * Added a note that the custom JavaScript is added to the global JavaScript tracker file and applies to all sites
+
+__5.0.3 - 2026-08-17__
+* Added code to harden the js save action
 
 __5.0.2__
 * Added missing license file